Update MDE and EdkModule packages for ICC build with /W4 /WX /Ox switches, for some latest modifications break it.
git-svn-id: https://edk2.svn.sourceforge.net/svnroot/edk2/trunk/edk2@2366 6f19259b-4bc3-4df7-8a09-765794883524

+++ MdePkg/Library/BaseLib/CheckSum.c
@@ -47,7 +47,7 @@ CalculateSum8 (
   ASSERT (Length <= (MAX_ADDRESS - ((UINTN) Buffer) + 1));
 
   for (Sum = 0, Count = 0; Count < Length; Count++) {
-    Sum = Sum + *(Buffer + Count);
+    Sum = (UINT8) (Sum + *(Buffer + Count));
   }
   
   return Sum;
@@ -126,7 +126,7 @@ CalculateSum16 (
 
 
   for (Sum = 0, Count = 0; Count < Length; Count++) {
-    Sum = Sum + *(Buffer + Count);
+    Sum = (UINT16) (Sum + *(Buffer + Count));
   }
   
   return Sum;
